Move loading and saving to separate component
