+++ src/app/paint/actions/mod.rs
@@ -6,6 +6,7 @@ pub use save_to_file::*;
 
 use crate::assets::Assets;
 use crate::backend::Image;
+use crate::iocomponent::IoComponent;
 use crate::paint_canvas::PaintCanvas;
 
 pub trait Action {
@@ -28,6 +29,7 @@ pub trait Action {
 pub struct ActionArgs<'a> {
    pub assets: &'a Assets,
    pub paint_canvas: &'a mut PaintCanvas,
+   pub iocomponent: &'a mut IoComponent,
 }
 
 fn _action_trait_must_be_object_safe(_action: Box<dyn Action>) {}
